+++ src/garage/tf/envs/__init__.py
@@ -1,5 +1,3 @@
 from garage.tf.envs.base import TfEnv
-from garage.tf.envs.parallel_vec_env_executor import ParallelVecEnvExecutor
-from garage.tf.envs.vec_env_executor import VecEnvExecutor
 
-__all__ = ['TfEnv', 'ParallelVecEnvExecutor', 'VecEnvExecutor']
+__all__ = ['TfEnv']
